updated file names and moved styling to styles folder. Successfully fixed volunteer color column on table
