--- src/app/globals.css
@@ -1,225 +1,0 @@
-/* General styles for mobile-first design */
-body {
-  background: linear-gradient(135deg, #f3e5f5, #e8f5e9); /* Soft gradient background */
-  color: #333; /* Darker text for better readability */
-  font-family: 'Arial', 'Helvetica', sans-serif;
-  margin: 0;
-  padding: 0;
-  text-align: center;
-}
-
-table {
-  width: 100%; /* Ensure the table takes full width on smaller screens */
-  border-collapse: collapse;
-  margin: 0 auto;
-  font-size: 14px; /* Reduce font size for smaller screens */
-  background-color: #ffffff; /* White background for the table */
-  border-radius: 8px; /* Rounded corners for the table */
-  overflow: hidden; /* Prevent content from overflowing */
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1); /* Subtle shadow for depth */
-}
-
-th {
-  background-color: #6a1b9a; /* Deep purple header background */
-  color: #ffffff; /* White text for headers */
-  padding: 12px; /* Add padding for better spacing */
-  text-align: center;
-  font-size: 14px;
-  font-weight: bold;
-}
-
-td {
-  border: 1px solid #e5e7eb; /* Light gray borders */
-  padding: 10px; /* Add padding for better spacing */
-  text-align: left;
-  font-size: 14px;
-}
-
-tr:nth-child(even) {
-  background-color: #f3f4f6; /* Light gray background for alternate rows */
-}
-
-tr:hover {
-  background-color: #d1c4e9; /* Light purple background on hover */
-}
-
-#table-container {
-  background-color: rgba(255, 255, 255, 0.8); /* Semi-transparent white background */
-  display: grid;
-  width: 100%;
-  overflow-x: auto; /* Allow horizontal scrolling for tables on smaller screens */
-  padding: 20px; /* Add padding for spacing */
-  border-radius: 10px;
-  margin: 20px auto;
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1); /* Subtle shadow for depth */
-}
-
-#newRowBtn {
-  margin: auto;
-  margin-top: 20px;
-  width: 90%; /* Adjust button width for smaller screens */
-  display: block;
-  padding: 0.7rem 1rem;
-  background-color: #4caf50; /* Green button background */
-  border: none;
-  border-radius: 8px; /* Rounded corners */
-  cursor: pointer;
-  font-weight: bold;
-  font-size: 16px; /* Adjust font size */
-  color: #ffffff; /* White text */
-  transition: background-color 0.3s ease, transform 0.2s ease;
-}
-
-#newRowBtn:hover {
-  background-color: #388e3c; /* Darker green on hover */
-  transform: scale(1.05); /* Slight zoom effect on hover */
-}
-
-.modal-content-hidden {
-  display: flex;
-  flex-direction: column; /* Stack elements vertically on smaller screens */
-  background-color: #ffffff; /* White background for the modal */
-  margin: auto;
-  padding: 20px;
-  border: 1px solid #e5e7eb; /* Light gray border */
-  border-radius: 15px;
-  width: 90%; /* Adjust modal width for smaller screens */
-  justify-content: space-between;
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1); /* Subtle shadow for depth */
-  z-index: 2;
-}
-
-.modal-hidden {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  gap: 0.4rem;
-  width: 90%; /* Adjust modal width for smaller screens */
-  padding: 1.3rem;
-  position: absolute;
-  top: 10%; /* Adjust modal position for smaller screens */
-  background-color: #ffffff; /* White background for the modal */
-  border-radius: 10px;
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1); /* Subtle shadow for depth */
-}
-
-.input {
-  padding: 10px;
-  border: 1px solid #d1d5db; /* Light gray border */
-  border-radius: 5px;
-  font-size: 14px;
-  width: 100%; /* Full width input */
-  margin-bottom: 10px; /* Add spacing between inputs */
-}
-
-#submit-button,
-#cancel-button {
-  width: 100%; /* Make buttons full-width on smaller screens */
-  margin-top: 10px;
-  padding: 0.7rem 1rem;
-  font-size: 16px; /* Adjust font size */
-  border: none;
-  border-radius: 8px; /* Rounded corners */
-  cursor: pointer;
-  font-weight: bold;
-  transition: background-color 0.3s ease, transform 0.2s ease;
-}
-
-#submit-button {
-  background-color: #4caf50; /* Green button background */
-  color: #ffffff; /* White text */
-}
-
-#submit-button:hover {
-  background-color: #388e3c; /* Darker green on hover */
-  transform: scale(1.05); /* Slight zoom effect on hover */
-}
-
-#cancel-button {
-  background-color: #f44336; /* Red button background */
-  color: #ffffff; /* White text */
-}
-
-#cancel-button:hover {
-  background-color: #d32f2f; /* Darker red on hover */
-  transform: scale(1.05); /* Slight zoom effect on hover */
-}
-
-.overlay {
-  position: fixed;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  width: 100vw;
-  height: 100vh;
-  background: rgba(0, 0, 0, 0.5); /* Dark overlay with transparency */
-  backdrop-filter: blur(5px); /* Add blur effect */
-  z-index: 1;
-}
-
-/* Media queries for smaller screens */
-@media screen and (max-width: 768px) {
-  #table-container {
-    width: 100%; /* Full width for smaller screens */
-  }
-
-  table {
-    font-size: 12px; /* Smaller font size for smaller screens */
-  }
-
-  th,
-  td {
-    padding: 8px; /* Adjust padding for smaller screens */
-  }
-
-  #newRowBtn {
-    width: 100%; /* Full width for the button */
-    font-size: 14px; /* Smaller font size */
-  }
-
-  .modal-hidden {
-    width: 90%; /* Adjust modal width */
-    top: 15%; /* Adjust modal position */
-  }
-}
-
-@media screen and (max-width: 480px) {
-  #table-container {
-    padding: 10px; /* Adjust padding for very small screens */
-  }
-
-  table {
-    font-size: 10px; /* Smaller font size for very small screens */
-  }
-
-  th,
-  td {
-    padding: 6px; /* Adjust padding further */
-  }
-
-  #newRowBtn {
-    font-size: 12px; /* Smaller font size */
-  }
-
-  .modal-hidden {
-    width: 95%; /* Adjust modal width */
-    top: 20%; /* Adjust modal position */
-  }
-}
-
-/* Fade-in animation */
-@keyframes fadeIn {
-  from {
-    opacity: 0;
-    transform: scale(0.95);
-  }
-  to {
-    opacity: 1;
-    transform: scale(1);
-  }
-}
-
-.animate-fade-in {
-  animation: fadeIn 0.3s ease-out;
-}
